# openapi_v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Invest Solo API",
    "description": "Quantitative fundamental analysis API for PEA and Global investment strategies.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/portfolio/": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get Portfolio",
        "description": "Return the full portfolio with live data, three-horizon scoring, and P&L.\n\nAll three horizon scores (score_lt, score_mt, score_st) are present on\nevery position. The ``horizon`` param only affects which signal column is\nused for the summary's signal_distribution bucket.",
        "operationId": "get_portfolio_api_portfolio__get",
        "parameters": [
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Horizon for signal_distribution in summary: long_term | medium_term | short_term",
              "default": "long_term",
              "title": "Horizon"
            },
            "description": "Horizon for signal_distribution in summary: long_term | medium_term | short_term"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/positions": {
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Add Position",
        "description": "Add a new position to the portfolio.",
        "operationId": "add_position_api_portfolio_positions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Add Position Api Portfolio Positions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/positions/{position_id}": {
      "put": {
        "tags": [
          "portfolio"
        ],
        "summary": "Update Position",
        "description": "Update an existing position.",
        "operationId": "update_position_api_portfolio_positions__position_id__put",
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Position Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePositionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Position Api Portfolio Positions  Position Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "portfolio"
        ],
        "summary": "Remove Position",
        "description": "Remove a position from the portfolio.",
        "operationId": "remove_position_api_portfolio_positions__position_id__delete",
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Position Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Remove Position Api Portfolio Positions  Position Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/refresh": {
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Refresh Portfolio",
        "description": "Force-refresh all live data for the portfolio.",
        "operationId": "refresh_portfolio_api_portfolio_refresh_post",
        "parameters": [
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Horizon for summary signal_distribution",
              "default": "long_term",
              "title": "Horizon"
            },
            "description": "Horizon for summary signal_distribution"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/company/{ticker}": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "Get Company Detail",
        "description": "Full company detail: metrics, three-horizon scoring, DCF, quality,\nrisk, momentum signals, analyst ratings, and PEA eligibility.",
        "operationId": "get_company_detail_api_company__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/company/{ticker}/horizons/{horizon}": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "Get Company Horizon",
        "description": "Single-horizon scoring view for a company.\n\nReturns the selected horizon's HorizonScoring block together with\nDCFValuation, QualitySignals, RiskSignals, and MomentumSignals so the\ncaller can fetch only the data relevant to a specific investment\ntimeframe without the full payload.\n\n``horizon`` must be one of: ``long_term``, ``medium_term``, ``short_term``.",
        "operationId": "get_company_horizon_api_company__ticker__horizons__horizon__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "horizon",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Horizon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Company Horizon Api Company  Ticker  Horizons  Horizon  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddPositionRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "minLength": 1,
            "title": "Ticker",
            "description": "Stock ticker symbol (e.g. TTE.PA)"
          },
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "description": "Number of shares"
          },
          "buy_price": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Buy Price",
            "description": "Average purchase price per share"
          },
          "buy_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Date",
            "description": "Date of purchase (YYYY-MM-DD)"
          },
          "account_type": {
            "type": "string",
            "pattern": "^(pea|pea_pme|cto)$",
            "title": "Account Type",
            "description": "Account type: pea, pea_pme, or cto",
            "default": "pea"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Free-text notes"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "quantity",
          "buy_price"
        ],
        "title": "AddPositionRequest",
        "description": "Request body for adding a new portfolio position."
      },
...
      "CompanyDetail": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "pea_eligible": {
            "type": "boolean",
            "title": "Pea Eligible",
            "default": false
          },
          "pea_pme_eligible": {
            "type": "boolean",
            "title": "Pea Pme Eligible",
            "default": false
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap"
          },
          "horizons": {
            "additionalProperties": {
              "$ref": "#/components/schemas/HorizonScoring"
            },
            "propertyNames": {
              "enum": [
                "long_term",
                "medium_term",
                "short_term"
              ]
            },
            "type": "object",
            "title": "Horizons",
            "description": "Horizon-specific scores, signals, and gate states"
          },
          "sub_scores": {
            "$ref": "#/components/schemas/SubScores",
            "description": "Six legacy M3 sub-scores"
          },
          "valuation": {
            "$ref": "#/components/schemas/DCFValuation",
            "description": "M6 three-scenario DCF block"
          },
          "quality": {
            "$ref": "#/components/schemas/QualitySignals",
            "description": "Piotroski / Altman / Graham / M5 quality"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskSignals",
            "description": "M5 risk signals + balance-sheet levers"
          },
          "momentum": {
            "$ref": "#/components/schemas/MomentumSignals",
            "description": "M9 price momentum and revision signals"
          },
          "metrics": {
            "$ref": "#/components/schemas/CompanyMetrics",
            "description": "Raw financial metrics (ratios, margins)"
          },
          "analyst_ratings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalystRatings"
              },
              {
                "type": "null"
              }
            ]
          },
          "data_completeness": {
            "type": "number",
            "title": "Data Completeness",
            "description": "Fraction of scoring inputs present 0-1",
            "default": 0.0
          },
          "data_source": {
            "type": "string",
            "title": "Data Source",
            "default": "yfinance"
          },
          "last_updated": {
            "type": "string",
            "title": "Last Updated",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "name",
          "horizons",
          "sub_scores",
          "valuation",
          "quality",
          "risk",
          "momentum",
          "metrics"
        ],
        "title": "CompanyDetail",
        "description": "Full company detail response (M10 schema).\n\nAll scoring is exposed through the three-horizon block; legacy\ncomposite_score / signal fields have been removed."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioPosition": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "account_type": {
            "type": "string",
            "title": "Account Type"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "buy_price": {
            "type": "number",
            "title": "Buy Price"
          },
          "buy_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Date"
          },
          "current_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Price"
          },
          "market_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Value"
          },
          "cost_basis": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Basis"
          },
          "gain_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gain Loss"
          },
          "gain_loss_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gain Loss Pct"
          },
          "weight_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Pct"
          },
          "pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe"
          },
          "pb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pb"
          },
          "ps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ps"
          },
          "roe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roe"
          },
          "roa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roa"
          },
          "roic": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roic"
          },
          "operating_margin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operating Margin"
          },
          "net_margin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Margin"
          },
          "revenue_growth": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenue Growth"
          },
          "div_yield": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Div Yield"
          },
          "score_lt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Lt",
            "description": "Long-term composite score 0-100"
          },
          "score_mt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Mt",
            "description": "Medium-term composite score 0-100"
          },
          "score_st": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score St",
            "description": "Short-term composite score 0-100"
          },
          "signal_lt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Lt",
            "description": "Long-term investment signal"
          },
          "signal_mt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Mt",
            "description": "Medium-term investment signal"
          },
          "signal_st": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal St",
            "description": "Short-term investment signal"
          },
          "piotroski_f": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Piotroski F"
          },
          "altman_z": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altman Z"
          },
          "graham_number": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graham Number"
          },
          "graham_mos": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graham Mos"
          },
          "dcf_mos_mid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dcf Mos Mid"
          },
          "pea_eligible": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pea Eligible"
          },
          "pea_pme_eligible": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pea Pme Eligible"
          },
          "analyst_rating": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Rating"
          },
          "analyst_target_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Target Price"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "forward_pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forward Pe"
          },
          "peg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peg"
          },
          "fifty_two_week_high_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fifty Two Week High Pct"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ticker",
          "account_type",
          "quantity",
          "buy_price"
        ],
        "title": "PortfolioPosition",
        "description": "A single enriched portfolio position with live data and three-horizon scoring."
      },
      "PortfolioResponse": {
        "properties": {
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PortfolioPosition"
            },
            "type": "array",
            "title": "Positions"
          },
          "summary": {
            "$ref": "#/components/schemas/PortfolioSummary"
          },
          "last_refreshed": {
            "type": "string",
            "title": "Last Refreshed"
          },
          "horizon": {
            "type": "string",
            "title": "Horizon",
            "description": "Horizon used for signal_distribution",
            "default": "long_term"
          }
        },
        "type": "object",
        "required": [
          "positions",
          "summary",
          "last_refreshed"
        ],
        "title": "PortfolioResponse",
        "description": "Full portfolio API response."
      },
...
      "UpdatePositionRequest": {
        "properties": {
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity",
            "description": "Updated number of shares"
          },
          "buy_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Price",
            "description": "Updated average purchase price"
          },
          "buy_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Date",
            "description": "Updated purchase date (YYYY-MM-DD)"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Updated notes"
          }
        },
        "type": "object",
        "title": "UpdatePositionRequest",
        "description": "Request body for updating an existing portfolio position."
      },
...
    }
  }
}